requirements.txt
anthropic>=0.40
pyyaml
requests
ddgs
torch
numpy
pandas
matplotlib
